feat: added block and ignore list feature
