+++ src/content/ignoreList.js
@@ -1,0 +1,8 @@
+export const ignoreList = [
+  "google.com",
+  "youtube.com",
+  "drive.google.com",
+  "docs.google.com",
+  "quran.com",
+  "islamqa.info",
+];
